refactor: extract ConfigurationFileManager from Program.cs
